+++ __init__.py
@@ -21,6 +21,7 @@
  ***************************************************************************/
  This script initializes the plugin, making it known to QGIS.
 """
+import imp
 import sys
 import os
 try:
@@ -28,22 +29,20 @@ try:
 except ImportError:
     pop_up_info = log = lambda x: x
 
-msg = ''
-
 sys.path.insert(
     0,
     os.path.join(os.path.dirname(os.path.realpath(__file__)), 'external')
 )
 
 try:
-    import sqlalchemy
-    import geoalchemy2
+    import sqlalchemy  # noqa
+    import geoalchemy2  # noqa
 except ImportError, e:
     pop_up_info("Error loading sqlalchemy or geoalchemy2 from "
                 "'external' subdirectory. error %s" % e.message)
 
 try:
-    import pyqtgraph
+    import pyqtgraph  # noqa
     log('Use local installation of pyqtgraph ')
 except Exception:
     # TODO: fix this error (which is the reason of this exception):
@@ -55,14 +54,20 @@ except Exception:
     pop_up_info(msg)
 
 try:
-    import lizard_connector
+    import lizard_connector  # noqa
 except ImportError, e:
     pop_up_info("Error loading lizard_connector from "
                 "'external' subdirectory. error %s" % e.message)
 
 try:
-    import netCDF4
-    msg += 'Use local installation of python netCDF4 library'
+    # Note: we're not importing it directly using the import statement because
+    # this will cause .pyd files to be loaded in dynamically. Because the
+    # loaded files are open in QGIS you can't delete them unless you close the
+    # program (at least with Windows), which is problematic when trying to
+    # update the plugin using the plugin manager (because it tries to delete
+    # the old plugin files). Real imports are postponed as long as possible.
+    imp.find_module('netCDF4')
+    log('Use local installation of python netCDF4 library')
 except ImportError:
     if os.name == 'nt':
         if sys.maxsize > 2**32:
@@ -71,29 +76,35 @@ except ImportError:
             sys.path.append(os.path.join(
                 os.path.dirname(os.path.realpath(__file__)),
                 'external', 'netCDF4-win64'))
-            import netCDF4
+            # import netCDF4
 
-            msg += 'Used netCDF4 library, provided with plugin. '
+            log('Used netCDF4 library, provided with plugin.')
         else:
             pop_up_info('Error: could not find netCDF4 installation. Change '
                         'to the 64-bit version of QGIS or try to install the '
                         'netCDF4 python libary yourself.')
-            netCDF4 = None
+            # netCDF4 = None
     else:
         pop_up_info('Error: could not find netCDF4 installation. Please '
                     'install python-netCDF4 package.')
-        netCDF4 = None
+        # netCDF4 = None
 
-if netCDF4 is not None:
-    msg += 'Python-netcdf version {python_netcdf}, netCDF4 version ' \
-           '{netcdf} and HDF5 version {netcdf}.'.format(
-               python_netcdf=netCDF4.__version__,
-               netcdf=netCDF4.__netcdf4libversion__,
-               hdf5=netCDF4.__hdf5libversion__)
-    log(msg)
+# if netCDF4 is not None:
+#     msg += 'Python-netcdf version {python_netcdf}, netCDF4 version ' \
+#            '{netcdf} and HDF5 version {netcdf}.'.format(
+#                python_netcdf=netCDF4.__version__,
+#                netcdf=netCDF4.__netcdf4libversion__,
+#                hdf5=netCDF4.__hdf5libversion__)
+#     log(msg)
 
 try:
-    import h5py
+    # Note: we're not importing it directly using the import statement because
+    # this will cause .pyd files to be loaded in dynamically. Because the
+    # loaded files are open in QGIS you can't delete them unless you close the
+    # program (at least with Windows), which is problematic when trying to
+    # update the plugin using the plugin manager (because it tries to delete
+    # the old plugin files). Real imports are postponed as long as possible.
+    imp.find_module('h5py')
     log("Using local h5py installation.")
 except ImportError as e:
     if os.name == 'nt':
@@ -101,7 +112,6 @@ except ImportError as e:
             sys.path.append(os.path.join(
                 os.path.dirname(os.path.realpath(__file__)),
                 'external', 'h5py-win64'))
-            import h5py
             log("Using h5py provided by plugin.")
         else:
             pop_up_info('Error: could not find h5py installation. Change '
@@ -111,6 +121,14 @@ except ImportError as e:
         pop_up_info('Error: could not find h5py installation. Please '
                     'install the h5py package manually.')
 
+try:
+    # Note: threedigrid depends on h5py and netCDF, so don't import it directly
+    # (see above).
+    imp.find_module('threedigrid') # noqa
+except ImportError as e:
+    pop_up_info("Error loading threedigrid from "
+                "'external' subdirectory. error %s" % e.message)
+
 
 # noinspection PyPep8Naming
 def classFactory(iface):  # pylint: disable=invalid-name
@@ -118,10 +136,7 @@ def classFactory(iface):  # pylint: disable=invalid-name
 
     :param iface: QgsInterface. A QGIS interface instance.
     """
-
     from .threedi_tools import ThreeDiTools
     from .utils.qlogging import setup_logging
-
     setup_logging(iface)
-
     return ThreeDiTools(iface)